apps/ai-engine/pyproject.toml
[project]
name = "smsa_ai_engine"
version = "0.1.0"
description = "Python AI engine for SMSA AI Assistant"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "aiohttp>=3.10.0",
  "xmltodict>=0.13.0",
  "langgraph>=0.1.0",
  "langchain-core>=0.2.0",
  "motor>=3.3.0",
  "esdk-obs-python>=3.24.0",
  "python-multipart>=0.0.6",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

